+++ tests/src/AlfBasicClassTest.php
@@ -1,7 +1,0 @@
-<?php
-
-test('one is one', function () {
-    //$obj = new AlfBasicClass();
-    //expect($obj->__debugInfo())->toBeArray();
-    expect(1)->toBe(1);
-});